+++ src/main/java/domain/card/deck/CardShuffleStrategy.java
@@ -1,5 +1,6 @@
-package domain.card;
+package domain.card.deck;
 
+import domain.card.Card;
 import java.util.List;
 
 @FunctionalInterface
